refactor: MVVM 패턴 디렉토리 구조로 변경
