+++ frontend/lib/data/datasources/kakao_local_datasource.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:frontend/models/place_model.dart';
+import 'package:frontend/data/models/place_model.dart';
 import 'package:http/http.dart' as http;
 
-class KakaoLocalService {
+class KakaoLocalDatasource {
   static const String _baseUrl = 'https://dapi.kakao.com/v2/local';
   static String get _apiKey => dotenv.env['KAKAO_REST_API_KEY'] ?? '';
