+++ frontend/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:frontend/presentation/screens/home_screen.dart';
+import 'package:frontend/presentation/views/home/home_screen.dart';
 import 'package:kakao_map_sdk/kakao_map_sdk.dart';
 
 void main() async {
