+++ frontend/lib/presentation/viewmodels/place_search_viewmodel.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../models/selected_place.dart';
+import 'package:frontend/data/models/place_model.dart';
 
 class PlaceNotifier extends StateNotifier<Map<String, SelectedPlace?>> {
   PlaceNotifier() : super({'start': null, 'end': null});
@@ -14,10 +14,7 @@ class PlaceNotifier extends StateNotifier<Map<String, SelectedPlace?>> {
 
   void swapPlaces() {
     final temp = state['start'];
-    state = {
-      'start': state['end'],
-      'end': temp,
-    };
+    state = {'start': state['end'], 'end': temp};
   }
 
   void clearPlaces() {
@@ -28,6 +25,7 @@ class PlaceNotifier extends StateNotifier<Map<String, SelectedPlace?>> {
   SelectedPlace? get endPlace => state['end'];
 }
 
-final placeProvider = StateNotifierProvider<PlaceNotifier, Map<String, SelectedPlace?>>((ref) {
-  return PlaceNotifier();
-});
+final placeProvider =
+    StateNotifierProvider<PlaceNotifier, Map<String, SelectedPlace?>>((ref) {
+      return PlaceNotifier();
+    });
